Enforce naming conventions across toolchain using clang-tidy.
This comes with one style clarification: for static, file-local
functions we had an inconsistent mix of prefixing their names with an
underscores vs not. The original intent was to have that prefix, but I
don't see a way to enforce it through clang-tidy, and it's hard to
find all those instances not having the prefix otherwise. So switching
over to not using an underscore prefix.