Fix C++ code generation for struct parameters passed as references.
Previously we could generate C++ moves for constructor arguments passed
as references. This is semantically incorrect, but seems to have had no
ill effects on code we generated. Since it could still confuse readers
and static analysis tools fix this to only generate moves if the
argument was passed as `copy`, but write out copies for all other
argument passing. This more clearly reflects what was already happening.