New `Optional` implementation.
So far the runtime used `std::optional` for representing optionals.
This commit replaces that with a custom `Optional` class that
encapsulates the implementation, and directly provides the semantics
we want. That means in particular that we can now remove most of the
former free functions,  and move their functionality directly into the
class. For example, the `value()` method now throws the expected HILTI
exception directly, without needing to remember to use a wrapper.  We
also rename some method to follow our coding conventions.

Behind the scenes, we continue to use `std::optional` for implementing
optionals, now through protected inheritance, similar like we do for
some other types as well.

This is all similar to how we already wrap other `std` types into
custom runtime types (e.g., vector).