Fix FPs in `FunctionParamVisitor` when parameters are shadowed.
The tracking of parameter use in `FunctionParamVisitor` relied on IDs to
correlate seen `expression::Name`s and parameter declarations. This
broke down when parameters where shadowed, e.g., we often end up
shadowing the `__cur` parameter.

With this patch we additionally check that the resolved `Declaration`
for an `Name` matches the parameter to take this into account.

Closes #2184.