Split out optimizer passes into their own compilation units.
This also introduces a pass registry so that each pass can hook itself
in, without the optimizer needing to have a compile-time dependency on
all of them.

The split is pretty mechanical for now, with some quick hacks to get
it working with minimal changes to the passes. More to come.