Split out optimizer passes into their own compilation units.
This also introduces a prelinnary pass registry so that each pass can
hook itself in the optimizer loop, without the optimizer needing to
have compile-time knowledge on all of them.

The split is pretty mechanical for now, with some quick hacks to get
it working with minimal changes to the passes; cleanup and
documentation coming in a subsequent commit.