Refactor optimizer.
This is a large change refactoring, streamlining, and speeding up of
the optimizer and its individual passes.

Conceptually, the core change is that we now have a well-defined
contract between optimizer and optimization passes: before running a
pass, the optimizer ensures that the AST is in a consistent,
fully-resolved state; and when a pass finishes, the pass tells the
optimizer what needs to be done to get the new AST back into that
state. See `optimizer::PassInfo` for the details.

We also substantially simplify the optimizer loop: it's no longer
nested across multiple phases & levels, and just runs all passes in
order until the AST stabilizes. Optionally, a individual pass maybe
itself repeated until stabilization, too, but that's the only nesting.

The passes themselves gain a new structure/API that clearly separates
between collection and mutation. We provide base visitor classes for
both kinds to derive from. The new structure no longer joins the
collection/mutation into a single visitor, which was hard to follow.

This new API enables the optimizer to track what parts of the AST were
changed by a pass, so that it can then re-resolve only the
corresponding subparts of the AST after each pass. We track changes
at the function-level so that we just need to recompute state for any
functions that have changed. The exception is global code/types, where
we re-resolve the whole module.

The optimizer's CFG state is centralized so all passes can share the
same, cached CFGs when possible, rather than each computing their own.
We use the function-tracking mentioned above to decide which CFGs need
re-computation.

The previous type/member/function passes got merged into single
dead-code-static pass, with joint collection and mutation visitors.

The constant-folder pass now re-uses the AST resolver's corresponding
functionality. No more two implementations of constant folding.

In debug builds, we do extensive checks that after a pass has
finished, and the AST state updated as prescribed, the AST is indeed
fully resolved again. We do this by recomputing all AST state from
scratch and comparing it against the current one. If there's a
difference, that's hard internal error and the compiler will abort.
This check is expensive and disabled in release builds. It can also be
disabled even in debug builds by setting a new CMake variable
HILTI_SKIP_EXPENSIVE_DEBUG_CHECKS.

This removes `HILTI_OPTIMIZER_PASSES` and
`HILTI_OPTIMIZER_ENABLE_CFG`. Neither seems that useful anymore, and
removing them simplifies the main optimizer loop. If we do still need
one of them, we can re-add later.