Put logic in place for update and validating AST state after optimizer passes.
We want the AST and its related state to be valid and fully resolved
after each optimizer pass. This puts two pieces in place towards that
goal:

1. Run a full resolver pass after each phase.

2. Afterwards, in debug builds, explicitly check everything is as we
   expect it to be. This check includes running the resolver once more
   to confirm that nothing changes anymore.

This is preliminary and currently slows down the optimizer quite a
bit: (1) is more that we need because fully resolving each time isn't
necessary in most cases; and (2) isn't giving us much additional value
right now because, well, we just resolved the whole thing. However,
the plan is to make (1) smarter going forward and try to update only
state that the pass may have touched. And then (2) should catch if we
screwed something up.