Reimplement `rt::currentExecutable`.
We previously used the `pathfind` library which isn't really necessary,
so Implement this functionality by hand.

The only major difference here is that instead of trying to infer the
the path to the current executable in OS-specific ways we now pass in
`argv[0]` which for all our uses allows to determine the location of the
current executable.