Add pass which performs block flattening.
A number of other rewrites currently depend on a simpler flow without
additional scopes. This patch implements such a pass which removes
nested scopes. This involves first renaming any identifiers which would
clash when lifted to the parent scope, and then merging the scope's
contents into its parent.

We also perform extra handling where an value aliased a value from the
removed scope. The previous behavior was that the alias would have
become dangling after the scope. Since this is something we explicitly
allow and diagnose at runtime we preserve that behavior by now pointing
the alias to other invalid data.

Closes #1425.