Fix non-determinism in computation of CFG data flow facts.
We were using unordered containers pretty heavily in the CFG which in
addition to the regular containers also requires equality. This seems to
play poorly with the pointers we put into these containers and introduce
non-determinism.

This patch replaces all unordered containers with their regular
counterparts which seems to take care of the issue.

Closes #2199.