Replace transform and filter utility functions with range adaptors.
These functions implemented functionality similar to what is available
from C++20 range adaptors, but instead of returning a view they would
eagerly collect into a vector, even if not needed.

This patch replaces uses with `view::transform` or `view::filter`.

The baseline change is due to a bug in the previous `transform`: if the
functor passed to it change the value type and we were transforming an
automatically sorting container like `std::set` we could accidentally
change the ordering. The issue here was here roughly that a set of `int`
has a different ordering than a set of these `int`s stringified due to
the difference between numeric and lexicographic order.