Enable clang-tidy enforce traditional ops.
This enforces that `&&` be preferred over `and` for all of the possible
operators. It's pretty confusing to mix them, and Spicy almost always
prefers the traditional versions, so seeing one-off keywords is weird.

I've also seen some compilers (MSVC, though we don't support it) disable
this by default in an extension. It's not worth the trouble for us.