Use $hlt instead of __hlt for internal namespace.
We prefix the namespace name with double underscore so it does not clash
with user-defined identifiers, but identifiers with leading underscores
are reserved by C++. In fact, `__hlt` is already defined for some ARM64
compilers where it refers to the `halt` instruction.

This patch replaces the leading `__` with `$`. While `$` is not valid in
identifiers in standard C++, it is allowed as an extension on MSVC, GCC
and Clang. In contrast to `__` it is not reserved though.

Closes #2205.