Extend HILTI group expression to support a local temporary.
This extends the HILTI `Grouping` expression (`(...)`) so that it
allows declaring a variable local to the group. To do that, one
prefixes the group's contained expression with a `local` declaration
inside `{...}`:

    ({local int i = 2;} i * i) # evaluates to 4

This isn't meant as a user-level feature but will be useful for
internal transformations. It's helpful for situations where an
expression is needed multiple times inside the group but must be
evaluated only once.

The HILTI syntax is pretty arbitrary and primarily meant to not
interfere with any code that's currently valid.

Due to the internal nature, we don't expose this in Spicy.