Uglify internal identifiers.
We prefixed the generated identifiers with double underscore so they do
not clash with user-defined identifiers, but identifiers with leading
underscores are reserved by C++. In fact, `__hlt` is already defined for
some ARM64 compilers where it refers to the `halt` instruction.

This patch replaces the leading `__` with a new reserved prefix
`hlt_internal` for identifiers in global scope, and else a `_`.

Closes #2205.