Fix ambiguous C++ constructors for generated struct.
This issue has been hidings for a while but was eventually bound to
come up: If a struct had just a single field or parameter, the
generated C++ struct would get a constructor that likewise receives a
single parameter. However, we also generated constructor copy and move
constructors, and if we're unlucky, then creating a struct with its
single argument would would become ambiguous with those constructors.
The second code example in https://github.com/zeek/spicy/issues/2197
taggers just that.

This fixes things by adding an additional tag parameter to all
constructors receiving struct fields/parameters. That disambiguates
them from the copy/move constructors; we should have done that a long
time ago already. The patch is pretty technical but mostly
straight-forward. The main challenge is finding all the places where
the tag needs to be added now.