Remove dependency of global pass registry on global logger.
Both these objects represent singletons which live in different
translation units. That means that the order in which they are
constructed is unspecified and we shouldn't rely on a particular one.

This patch removes a user of the logger to raise an internal error with
an assertion to break that dependency.