Add rvalue overloads for `Result` accessors.
A somewhat typical use case of results is to get them from some function
and then immediately derference the returned value and bind its value to
a new value (i.e., not a reference). The binding to a new value prevents
us introducing a reference to a temporary which will go out of scope and
the end of the statement.

This is incorrectly diagnosed by Coverity as a unnecessary copy
(`AUTO_CAUSES_COPY`), but also exposes an API which makes it easy to
create dangling references. This patch introduces rvalue overloads for
`Result` accessors so we can directly obtain a value inside a `Result`
without having to go through a copy of a reference. With that we also
catch incorrect uses of temporary `Result` values which we also fix.