Separate lookahead and size benchmarks.
Since the two benchmarks used the same `Inner`, they would get the same
interprocedural optimizations. This doesn't give the full picture about
how something changes, so separate them so that we can get a better
picture.

This definitely makes the benchmarks more "micro," but I believe that we
see more people using *either* lookahead *or* size, not both on the same
unit.