Factor out logic & state for CFG caching into separate class.
This encapsulates the functionality, making it reusable; and will also
allow us to later pass the cache around to different parts of the
compiler pipeline.

While the internal structure of the cache is slightly different to the
previous implementation inside the optimizer, there's no functionality
change to the optimizer except for one fix: we were not correctly
discarding all affected block CFGs when invalidating a module. We do
now, with the result that we're computing some more CFGs overall,
costing us about 1-2% in performance.