Replace assignments with their RHS, not remove.
After 5197e79, we replace some declarations with the RHS of the
assignment so that we don't remove method calls. The same problem applies
to assignments generally, so add cases for those.

This also makes some code stick around (particularly when paired with
constant propagation), so we need more aggressive removal of expressions
whose result is unused and who don't have side effects.