Take operand constness into account in CFG construction.
When computing the CFG data access information we would previously
assume that any operator use would both read and write its operands,
e.g., we would assume that an equality check could write its operands.
Due to how we resolve operators of non-const operands this does reflect
what we see in many ASTs, but does not take into account when we
actually use an operand as const.

This patch adds handling of operators so that we can avoid recording
writes when the operand is used as const.